imp - Finalized list implementation for ID3v2
We've adjusted values as necessary.

---

We've adjusted the struct as necessary to be able to get correct info from the ID3v2 tag extras and texts. This is because of the difference of C-style ints vs. C#-style ints.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1

+++ BassBoom.Native/Interop/Analysis/NativeMetadata.cs
@@ -73,8 +73,8 @@ namespace BassBoom.Native.Interop.Analysis
     public unsafe struct mpg123_string
     {
         internal char* p;
-        internal int size;
-        internal int fill;
+        internal IntPtr size;
+        internal IntPtr fill;
     }
 
     [StructLayout(LayoutKind.Sequential)]
@@ -94,7 +94,7 @@ namespace BassBoom.Native.Interop.Analysis
         internal char type;
         internal mpg123_string description;
         internal mpg123_string mime_type;
-        internal int size;
+        internal IntPtr size;
         internal char* data;
     }
